+++ Plugins/ProcGen/Source/NexusProcGen/Private/Cell/NCellProxy.cpp
@@ -42,7 +42,6 @@ ANCellProxy::ANCellProxy(const FObjectInitializer& ObjectInitializer)
 
 ANCellProxy* ANCellProxy::CreateInstance(UWorld* World, UNCell* Cell, const FVector& Location, const FRotator& Rotation, bool bPreLoadLevel)
 {
-	// TODO: Make const?
 	FActorSpawnParameters SpawnInfo;
 	
 	SpawnInfo.ObjectFlags |= RF_Transient;