+++ Plugins/ProcGen/Source/NexusProcGen/Public/Generation/NProcGenOperationSharedContext.h
@@ -9,6 +9,7 @@ class FNProcGenOperationSharedContext
 public:
 	UWorld* TargetWorld;
 	TArray<ANCellProxy*> CreatedProxies;
+	TArray<AActor*> CreatedActors;
 
 	explicit FNProcGenOperationSharedContext(UWorld* OutputWorld);
 };