+++ Plugins/ProcGen/Source/NexusProcGenEditor/Private/NProcGenEditorSubsystem.cpp
@@ -15,7 +15,7 @@ void UNProcGenEditorSubsystem::StartOperation(UNProcGenOperation* Operation)
 		FName LastGenerationKey = Component->GetAndResetGenerationOperationKey();
 		if (LastGenerationKey != NAME_None)
 		{
-			ClearGeneratedProxies(LastGenerationKey);
+			ClearGenerated(LastGenerationKey);
 		}
 	}
 	
@@ -26,6 +26,8 @@ void UNProcGenEditorSubsystem::StartOperation(UNProcGenOperation* Operation)
 void UNProcGenEditorSubsystem::OnOperationFinished(UNProcGenOperation* Operation, TSharedRef<FNProcGenOperationSharedContext> SharedContext)
 {
 	KnownProxies.Add(Operation->GetFName(), TArray<ANCellProxy*>(SharedContext->CreatedProxies)); // TODO : Check this copies?
+	KnownActors.Add(Operation->GetFName(), TArray<AActor*>(SharedContext->CreatedActors)); // TODO : Check this copies?
+	
 	KnownOperations.Remove(Operation);
 }
 
@@ -34,7 +36,7 @@ void UNProcGenEditorSubsystem::OnOperationDestroyed(UNProcGenOperation* Operatio
 	KnownOperations.Remove(Operation);
 }
 
-void UNProcGenEditorSubsystem::ClearAllGeneratedProxies()
+void UNProcGenEditorSubsystem::ClearAllGenerated()
 {
 	if (KnownProxies.Num() > 0)
 	{
@@ -45,6 +47,18 @@ void UNProcGenEditorSubsystem::ClearAllGeneratedProxies()
 			ClearGeneratedProxies(Key);
 		}
 		KnownProxies.Empty();
+	}
+	
+	
+	if (KnownActors.Num() > 0)
+	{
+		TArray<FName> KnownKeys;
+		KnownActors.GetKeys(KnownKeys);
+		for (auto Key : KnownKeys)
+		{
+			ClearGeneratedActors(Key);
+		}
+		KnownActors.Empty();
 	}
 }
 
@@ -60,6 +74,20 @@ void UNProcGenEditorSubsystem::ClearGeneratedProxies(const FName& Key)
 			ProxiesArray[i]->Destroy(true, false);
 		}
 		KnownProxies.Remove(Key);
+	}
+}
+
+void UNProcGenEditorSubsystem::ClearGeneratedActors(const FName& Key)
+{
+	if (KnownActors.Num() > 0 && KnownActors.Contains(Key))
+	{
+		TArray<AActor*> ActorsArray = *KnownActors.Find(Key);
+		const int FoundCount = ActorsArray.Num();
+		for (int i = 0; i < FoundCount; i++)
+		{
+			ActorsArray[i]->Destroy(true, false);
+		}
+		KnownActors.Remove(Key);
 	}
 }